Add option for binding to address #626
