+++ src/local.c
@@ -131,19 +131,6 @@ int setnonblocking(int fd)
 
 #endif
 
-#ifdef SET_INTERFACE
-int setinterface(int socket_fd, const char *interface_name)
-{
-    struct ifreq interface;
-    memset(&interface, 0, sizeof(interface));
-    strncpy(interface.ifr_name, interface_name, IFNAMSIZ);
-    int res = setsockopt(socket_fd, SOL_SOCKET, SO_BINDTODEVICE, &interface,
-                         sizeof(struct ifreq));
-    return res;
-}
-
-#endif
-
 int create_and_bind(const char *addr, const char *port)
 {
     struct addrinfo hints;
